+++ apps/__init__.py
@@ -10,21 +10,17 @@ from flask_login import LoginManager
 from flask_sqlalchemy import SQLAlchemy
 from importlib import import_module
 
-
 db = SQLAlchemy()
 login_manager = LoginManager()
-
 
 def register_extensions(app):
     db.init_app(app)
     login_manager.init_app(app)
 
-
 def register_blueprints(app):
-    for module_name in ('authentication', 'home', 'tables', 'dyn_api'):
+    for module_name in ('authentication', 'home', 'dyn_dt'):
         module = import_module('apps.{}.routes'.format(module_name))
         app.register_blueprint(module.blueprint)
-
 
 def configure_database(app):